Use package to organise files
