+++ src/main/java/nugget/Command.java
@@ -1,3 +1,5 @@
+package nugget;
+
 public interface Command {
     void execute(TaskList tasks, Ui ui, Storage storage) throws NuggetException;
 }
