+++ src/main/java/nugget/Deadline.java
@@ -1,3 +1,5 @@
+package nugget;
+
 import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
 import java.time.format.DateTimeParseException;
