+++ src/main/java/nugget/EmptyDescriptionException.java
@@ -1,3 +1,5 @@
+package nugget;
+
 public class EmptyDescriptionException extends NuggetException {
     public EmptyDescriptionException() {
         super("OOPS!!! The description of a task cannot be empty.");
