+++ src/main/java/nugget/InvalidTaskNumberException.java
@@ -1,3 +1,5 @@
+package nugget;
+
 public class InvalidTaskNumberException extends NuggetException {
     public InvalidTaskNumberException() {
         super("OOPS!!! The task number you provided is invalid.");
