+++ src/main/java/nugget/NuggetException.java
@@ -1,3 +1,5 @@
+package nugget;
+
 public class NuggetException extends Exception {
     public NuggetException(String message) {
         super(message);
