+++ src/main/java/nugget/Task.java
@@ -1,3 +1,5 @@
+package nugget;
+
 public class Task {
     protected String description;
     protected boolean isDone;
