+++ src/main/java/nugget/UnknownCommandException.java
@@ -1,3 +1,7 @@
+package nugget;
+
+import nugget.NuggetException;
+
 public class UnknownCommandException extends NuggetException {
     public UnknownCommandException() {
         super("OOPS!!! Please enter valid commands!");
