+++ text-ui-test/EXPECTED.TXT
@@ -1,5 +1,5 @@
 ________________________________________
-Hello! I'm Nugget
+Hello! I'm nugget.Nugget
 What can I do for you?
 ________________________________________
 ________________________________________
